first change switching from in memory to persistence with typeorm and postgresql
