+++ src/tasks/dto/update-task-status.dto.ts
@@ -1,5 +1,5 @@
 import { IsEnum } from 'class-validator';
-import { TaskStatus } from '../task.model';
+import { TaskStatus } from '../task-status.enum';
 
 export class UpdateTaskStatusDto {
   @IsEnum(TaskStatus)
