+++ src/tasks/task.entity.ts
@@ -1,0 +1,17 @@
+import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
+import { TaskStatus } from './task-status.enum';
+
+@Entity()
+export class Task {
+  @PrimaryGeneratedColumn('uuid') // automatically generate id & set as primary column
+  id: string;
+
+  @Column()
+  title: string;
+
+  @Column()
+  description: string;
+
+  @Column()
+  status: TaskStatus;
+}
