+++ src/tasks/tasks.module.ts
@@ -1,8 +1,11 @@
 import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
 import { TasksController } from './tasks.controller';
+import { TasksRepository } from './tasks.repository';
 import { TasksService } from './tasks.service';
 
 @Module({
+  imports: [TypeOrmModule.forFeature([TasksRepository])],
   controllers: [TasksController],
   providers: [TasksService],
 })
