+++ src/tasks/tasks.repository.ts
@@ -1,0 +1,5 @@
+import { EntityRepository, Repository } from 'typeorm';
+import { Task } from './task.entity';
+
+@EntityRepository(Task)
+export class TasksRepository extends Repository<Task> {}
